+++ apibuilder4elastic/test/trafficMonitorAPI/asAdmin/http/test_search_custom_restrictionPermissions.js
@@ -1,0 +1,96 @@
+const { expect } = require('chai');
+const { startApiBuilder, stopApiBuilder, requestAsync, sendToElasticsearch, getRandomInt } = require('../../../_base');
+const path = require('path');
+const fs = require('fs');
+const nock = require('nock');
+const envLoader = require('dotenv');
+
+describe('Endpoints', function () {
+	this.timeout(30000);
+	let server;
+	let auth;
+	const indexName = `apigw-traffic-summary-search_custom_restricted_perm_test_${getRandomInt(9999)}`;
+
+	afterEach(() => {
+		nock.cleanAll();
+	});
+
+	/**
+	 * Start API Builder.
+	 */
+	before(() => {
+		return new Promise(function(resolve, reject){
+			const envFilePath = path.join(__dirname, '../../../.env');
+			if (fs.existsSync(envFilePath)) {
+				envLoader.config({ path: envFilePath });
+			}
+			// Special permissions configured required to get unrestricted traffic access
+			process.env.UNRESTRICTED_PERMISSIONS = "logs,mgmt";
+			server = startApiBuilder();
+			server.apibuilder.config.testElasticIndex = indexName;
+			elasticConfig = server.apibuilder.config.pluginConfig['@axway-api-builder-ext/api-builder-plugin-fn-elasticsearch'].elastic;
+			server.started
+			.then(() => {
+				const entryset = require('../../../documents/http/search_count_documents');
+				sendToElasticsearch(elasticConfig, indexName, 'traffic-summary/index_template.json', entryset)
+				.then(() => {
+					resolve();
+				})
+				.catch(err => reject(err));
+			});
+		});
+	});
+
+	/**
+	 * Stop API Builder after the tests.
+	 */
+	 after(() => {
+		stopApiBuilder(server);
+		delete process.env.UNRESTRICTED_PERMISSIONS;
+	});
+
+	describe('Search', () => {
+		it('[RestrictedPermissions-0001] Execute a search using a custom restricted permission configuration', () => {
+			// Permissions returned include mgmt & logs
+			nock('https://mocked-api-gateway:8090').get('/api/rbac/currentuser').reply(200, { "result": "david" });
+			nock('https://mocked-api-gateway:8090').get('/api/topology').reply(200, { result: {} });
+			nock('https://mocked-api-gateway:8090').get('/api/rbac/permissions/currentuser').replyWithFile(200, './test/mockedReplies/apigateway/permissionLogsMgmt.json');
+			return requestAsync({
+				method: 'GET',
+				uri: `http://localhost:${server.apibuilder.port}/api/elk/v1/api/router/service/instance-1/ops/search?count=20`,
+				headers: {
+					'cookie': 'VIDUSR=Search-Count-0001-DAVID-1597468226-Z+qdRW4rGZnwzQ==', 
+					'csrf-token': '04F9F07E59F588CDE469FC367A12ED3A4B845FDA9A9AE2D9A77686823067CDDC'
+				},
+				json: true
+			}).then(({ response, body }) => {
+				expect(response.statusCode).to.equal(200);
+				expect(body).to.be.an('Object');
+				expect(body).to.have.property('data');
+				expect(body.data).to.have.lengthOf(13); // Expect the entire result set
+			});
+		});
+
+		it('[RestrictedPermissions-0002] Custom restriction permission configuration - User has with monitoring permission a restricted access', () => {
+			// Configured permissions: mgmt,logs - Not enough for user rene having permission monitoring
+			nock('https://mocked-api-gateway:8090').get('/api/rbac/currentuser').reply(200, { "result": "rene" });
+			nock('https://mocked-api-gateway:8090').get('/api/rbac/permissions/currentuser').replyWithFile(200, './test/mockedReplies/apigateway/permissionMonitoring.json');
+			nock('https://mocked-api-gateway:8075').get(`/api/portal/v1.3/users?field=loginName&op=eq&value=rene&field=enabled&op=eq&value=enabled`).replyWithFile(200, './test/mockedReplies/apimanager/apiManagerUserRene.json');		
+			nock('https://mocked-api-gateway:8075').get(`/api/portal/v1.3/organizations/2bfaa1c2-49ab-4059-832d-MAX`).replyWithFile(200, './test/mockedReplies/apimanager/organizationMax.json');
+			return requestAsync({
+				method: 'GET',
+				uri: `http://localhost:${server.apibuilder.port}/api/elk/v1/api/router/service/instance-1/ops/search?count=20`,
+				headers: {
+					'cookie': 'VIDUSR=Search-Count-0001-DAVID-1597468226-Z+qdRW4rGZnwzQ==', 
+					'csrf-token': '04F9F07E59F588CDE469FC367A12ED3A4B845FDA9A9AE2D9A77686823067CDDC'
+				},
+				json: true
+			}).then(({ response, body }) => {
+				expect(response.statusCode).to.equal(200);
+				expect(body).to.be.an('Object');
+				expect(body).to.have.property('data');
+				expect(body.data).to.have.lengthOf(2); // Expect only two results as the user is restricted
+			});
+		});
+	});
+});